pyproject.toml
[tool.poetry]
name = "health-recommendation-agent"
version = "0.1.0"
description = "Simple health recommendation agent"
authors = ["Aman Singh"]

[tool.poetry.dependencies]
python = ">=3.11.0,<=3.11.2"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.19"
unstructured = "^0.15.0"
crewai = "^0.41.1"
pymupdf = "^1.24.9"
flask = "^3.1.0"
openai = "^1.58.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"